+++ AgeOfEmpiresTaller/VistaSrc/FactoryDibujo.cpp
@@ -17,15 +17,15 @@ FactoryDibujo::FactoryDibujo(SDL_Renderer* gRenderer) {
 	this->cantidad_de_dibujos=this->dibujos.size();
 	this->dibujo_actual=1;
 }
-bool FactoryDibujo::crear_dibujo_personaje(string path,size_t cantidad_de_movimientos,vector<int> cantidad_de_imagenes,std::vector<std::vector<int> >parametros_de_imagen,int fps,int velocidad){
+bool FactoryDibujo::crear_dibujo_personaje(string path,int cantidad_de_movimientos,vector<int> cantidad_de_imagenes,std::vector<std::vector<int> >parametros_de_imagen,int fps,int velocidad){
 	shared_ptr<DibujoPersonaje> dibujo_nuevo = shared_ptr<DibujoPersonaje>(new DibujoPersonaje());
 	if (!dibujo_nuevo->cargar_archivo(path,gRenderer)) {
 		printf("Failed to load image!\n");
 	} else {
 		dibujo_nuevo->set_cantidad_de_movimientos(cantidad_de_movimientos);
-		for(size_t i=0;i<cantidad_de_movimientos;i++){
+		for(int i=0;i<cantidad_de_movimientos;i++){
 			dibujo_nuevo->set_cantidad_de_imagenes(i,cantidad_de_imagenes[i]);
-			for (size_t j = 0; j < cantidad_de_imagenes[i]; j++) {
+			for (int j = 0; j < cantidad_de_imagenes[i]; j++) {
 						dibujo_nuevo->set_imagen(i,j, parametros_de_imagen[i][X], parametros_de_imagen[i][Y],parametros_de_imagen[i][ANCHO], parametros_de_imagen[i][ALTO]);
 			}
 		}
@@ -36,7 +36,7 @@ bool FactoryDibujo::crear_dibujo_personaje(string path,size_t cantidad_de_movimi
 	}
 	return false;
 }
-bool FactoryDibujo::crear_dibujo_animado(string path,size_t cantidad_de_imagenes,std::vector<std::vector<int> >parametros_de_imagen,int fps){
+bool FactoryDibujo::crear_dibujo_animado(string path,int cantidad_de_imagenes,std::vector<std::vector<int> >parametros_de_imagen,int fps){
 	shared_ptr<DibujoAnimado> dibujo_nuevo = shared_ptr<DibujoAnimado>(new DibujoAnimado());
 	if (!dibujo_nuevo->cargar_archivo(path, gRenderer)) {
 		printf("Failed to load image!\n");
@@ -62,7 +62,7 @@ bool FactoryDibujo::crear_dibujo_estatico(string path, vector<int>parametros_de_
 	}
 	return false;
 }
-size_t FactoryDibujo::ultimo_dibujo(){
+int FactoryDibujo::ultimo_dibujo(){
 	return (dibujo_actual-1);
 }
 
@@ -74,7 +74,7 @@ void FactoryDibujo::set_dibujo(shared_ptr<Dibujo> dibujo){
 	dibujo_actual++;
 }
 
-shared_ptr<Dibujo> FactoryDibujo::get_dibujo(size_t n_dibujo){
+shared_ptr<Dibujo> FactoryDibujo::get_dibujo(int n_dibujo){
 	if(n_dibujo>=this->dibujo_actual){
 		return NULL;
 	}
